Fix false positive in `supports` (#14)
Fix the case when e.g. `supports('deriveKey', 'PBKDF2', 'HKDF')` returns
true because the length parameter from the "get key length" operation of
HKDF (which returns null, which causes the "deriveKey" operation of
PBKDF2 to throw) is not available.
(The same is true for other combinations of PBKDF2 and HKDF.)